requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-phone-field==1.8.1
django-phonenumber-field==6.4.0
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2
